+++ src/app/models/enum/user.enum.ts
@@ -1,5 +1,16 @@
 export enum userRolesEnum {
   ADMIN,
   ADVERTISER,
-  PUBLISHER
-};
+  PUBLISHER,
+}
+
+export enum reportType {
+  CAMPAIGNS = 'campaigns',
+  SITES = 'sites',
+}
+
+export enum reportState {
+  PREPARING = 'preparing',
+  READY = 'ready',
+  DELETED = 'deleted',
+}